Add support for using virtualenv #181
It's also possible to use virtualenv, instead of Docker, to get a
development environment up and running. This commit adds documentation
to the README and updates the requirements.txt on how to do this.

* README.md
Add instructions

* requirements.txt
(psycopg2-binary): upgrade to 2.8.6
(django-storages): Use instead of chris' custom branch which is not
 currently maintained.
(ipython): Useful for using shell_plus in `python manage.py shell_plus`